@@ -16,45 +16,45 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^13.0.2",
+    "@angular/animations": "^13.1.0",
     "@angular/cdk": "^13.0.0",
-    "@angular/common": "^13.0.3",
+    "@angular/common": "^13.1.1",
     "@angular/compiler": "^13.0.0",
     "@angular/core": "^13.0.1",
-    "@angular/forms": "^13.0.3",
-    "@angular/material": "^13.0.0",
-    "@angular/platform-browser": "^13.0.3",
+    "@angular/forms": "^13.1.1",
+    "@angular/material": "^13.1.1",
+    "@angular/platform-browser": "^13.1.1",
     "@angular/platform-browser-dynamic": "^13.0.2",
-    "@angular/router": "^13.0.3",
-    "autoprefixer": "^10.4.0",
+    "@angular/router": "^13.1.1",
     "ng-click-outside": "^9.0.1",
-    "ngx-cookie-service": "^13.0.1",
-    "ngx-mask": "^13.0.0",
+    "ngx-cookie-service": "^13.1.2",
+    "ngx-mask": "^13.0.1",
     "rxjs": "^7.4.0",
     "tslib": "^2.3.1",
     "uuid": "^8.3.2",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/architect": "^0.1300.1",
-    "@angular-devkit/build-angular": "^13.0.1",
+    "@angular-devkit/architect": "^0.1301.1",
+    "@angular-devkit/build-angular": "^13.1.2",
     "@angular-eslint/builder": "^13.0.1",
     "@angular-eslint/eslint-plugin": "^13.0.1",
     "@angular-eslint/eslint-plugin-template": "^13.0.1",
     "@angular-eslint/schematics": "^13.0.1",
     "@angular-eslint/template-parser": "^13.0.1",
-    "@angular/cli": "^13.0.1",
-    "@angular/compiler-cli": "^13.0.3",
+    "@angular/cli": "^13.1.2",
+    "@angular/compiler-cli": "^13.1.1",
     "@compodoc/compodoc": "^1.1.16",
     "@strut-software/eslint-config": "^1.6.1",
     "@strut-software/stylelint-config": "^1.1.1",
     "@types/jasmine": "^3.10.2",
-    "@types/node": "^16.11.11",
+    "@types/node": "^16.11.7",
     "@types/uuid": "^8.3.3",
-    "@typescript-eslint/eslint-plugin": "5.5.0",
-    "@typescript-eslint/parser": "5.5.0",
+    "@typescript-eslint/eslint-plugin": "5.8.0",
+    "@typescript-eslint/parser": "5.8.0",
+    "autoprefixer": "^10.4.0",
     "cross-env": "^7.0.3",
-    "eslint": "8.4.0",
+    "eslint": "8.5.0",
     "husky": "^7.0.4",
     "jasmine-core": "~3.10.1",
     "jasmine-spec-reporter": "~7.0.0",
@@ -63,14 +63,15 @@
     "karma-coverage": "~2.1.0",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.7.0",
-    "ng-mocks": "^12.5.0",
-    "pa11y-ci": "3.0.0",
-    "postcss-cli": "^9.0.2",
+    "ng-mocks": "^12.5.1",
+    "pa11y-ci": "3.0.1",
+    "postcss": "^8.4.5",
+    "postcss-cli": "^9.1.0",
     "sassdoc": "^2.7.3",
     "stylelint": "^13.13.1",
     "stylelint-config-standard": "^22.0.0",
     "stylelint-scss": "^3.21.0",
-    "tailwindcss": "^2.2.19",
+    "tailwindcss": "^3.0.7",
     "ts-node": "~10.4.0",
     "typescript": "~4.4.4"
   }
