@@ -2,7 +2,7 @@ import { HttpClientTestingModule, HttpTestingController } from '@angular/common/
 import { TestBed } from '@angular/core/testing';
 import { environment } from 'src/environments/environment';
 // eslint-disable-next-line max-len
-import { ForwardPreviousStationsDocument, StationDocuments, UserType, DocumentStationInformation, StandardStringJSON } from 'src/models';
+import { ForwardPreviousStationsDocument, StationDocuments, UserType, DocumentStationInformation, StandardStringJSON, DocumentAnswer, QuestionFieldType, DocumentName, StationRosterMember, Question, DocumentAutoFlow } from 'src/models';
 import { DocumentService } from './document.service';
 
 const MICROSERVICE_PATH = '/documentservice/api/document';
@@ -76,7 +76,7 @@ describe('DocumentService', () => {
           totalDocuments: 2
         }
       ],
-      followingStations: [
+      nextStations: [
         {
           rithmId: '852-963-741',
           name: 'Follow station 1',
@@ -139,25 +139,219 @@ describe('DocumentService', () => {
   });
 
   it('Should return the update of the new document name', () => {
-    const documentName: StandardStringJSON = {
-      data: 'Almond Flour'
+    const documentName = 'Almond Flour';
+    const expectDocumentName: StandardStringJSON = {
+      data: documentName
     };
 
     service.updateDocumentName(documentId, documentName)
       .subscribe((newDocumentName) => {
         expect(newDocumentName).toEqual(documentName);
       });
+
+    const req = httpTestingController.expectOne(`${environment.baseApiUrl}${MICROSERVICE_PATH}/name?rithmId=${documentId}`);
+    expect(req.request.method).toEqual('PUT');
+
+    expect(req.request.body).toEqual(expectDocumentName);
+    req.flush(expectDocumentName);
+    httpTestingController.verify();
   });
 
   it('should return document name', () => {
-    const documentName: StandardStringJSON = {
-      data: 'Metroid Dread'
+    const documentName: DocumentName = {
+      baseName: 'Metroid Dread',
+      appendedName: ''
     };
+
 
     service.getDocumentName(documentId)
       .subscribe((response) => {
         expect(response).toEqual(documentName);
       });
+
+    const req = httpTestingController.expectOne(`${environment.baseApiUrl}${MICROSERVICE_PATH}/name?documentRithmId=${documentId}`);
+    expect(req.request.method).toEqual('GET');
+    expect(req.request.params.get('documentRithmId')).toBe(documentId);
+    req.flush(documentName);
+    httpTestingController.verify();
+
   });
 
+
+  it('should make request to save document answer', () => {
+    const expectedAnswers: DocumentAnswer[] = [{
+      questionRithmId: 'Dev 1',
+      documentRithmId: '123-654-789',
+      stationRithmId: '741-951-753',
+      value: 'Answer Dev',
+      file: 'dev.txt',
+      filename: 'dev',
+      type: QuestionFieldType.Email,
+      rithmId: '789-321-456',
+      questionUpdated: true,
+    },
+    {
+      questionRithmId: 'Dev 2',
+      documentRithmId: '123-654-789-856',
+      stationRithmId: '741-951-753-741',
+      value: 'Answer Dev2',
+      file: 'dev2.txt',
+      filename: 'dev2',
+      type: QuestionFieldType.City,
+      rithmId: '789-321-456-789',
+      questionUpdated: false,
+    }];
+
+    service.saveDocumentAnswer(documentId, expectedAnswers)
+      .subscribe((response) => {
+        expect(response).toEqual(expectedAnswers);
+      });
+
+    // eslint-disable-next-line max-len
+    const req = httpTestingController.expectOne(`${environment.baseApiUrl}${MICROSERVICE_PATH}/answers?documentRithmId=${documentId}`);
+    expect(req.request.method).toEqual('POST');
+    expect(req.request.body).toEqual(expectedAnswers);
+
+    req.flush(expectedAnswers);
+    httpTestingController.verify();
+  });
+
+  it('should return updated date from a specific document', () => {
+    const expectedResponse: StandardStringJSON = {
+      data: '2021-12-09T17:26:47.3506612Z'
+    };
+
+    service.getLastUpdated(documentId)
+      .subscribe((response) => {
+        expect(response).toEqual(expectedResponse.data);
+      });
+
+    const req = httpTestingController.expectOne(`${environment.baseApiUrl}${MICROSERVICE_PATH}/last-updated?documentRithmId=${documentId}`);
+    expect(req.request.method).toEqual('GET');
+    expect(req.request.params.get('documentRithmId')).toBe(documentId);
+    req.flush(expectedResponse);
+    httpTestingController.verify();
+  });
+
+  it('should return held time in station for document', () => {
+    const expectedResponse: StandardStringJSON = {
+      data: '2021-12-09T17:26:47.3506612Z'
+    };
+
+    service.getDocumentTimeInStation(documentId, stationId)
+      .subscribe((documentTimeInStation) => {
+        expect(documentTimeInStation).toEqual(expectedResponse.data);
+      });
+
+    // eslint-disable-next-line max-len
+    const req = httpTestingController.expectOne(`${environment.baseApiUrl}${MICROSERVICE_PATH}/flowed-time?documentRithmId=${documentId}&stationRithmId=${stationId}`);
+    expect(req.request.method).toEqual('GET');
+    expect(req.request.params.get('documentRithmId')).toBe(documentId);
+    expect(req.request.params.get('stationRithmId')).toBe(stationId);
+    req.flush(expectedResponse);
+    httpTestingController.verify();
+  });
+
+  it('should return previous questions', () => {
+    const expectPreviousQuestions: Question[] = [
+      {
+        rithmId: '',
+        questionType: QuestionFieldType.City,
+        prompt: 'string',
+        isPrivate: true,
+        isEncrypted: true,
+        isReadOnly: true,
+        isRequired: true,
+        possibleAnswers: [
+          {
+            text: 'string',
+            default: true
+          }
+        ],
+        answer: {
+          questionRithmId: 'string',
+          referAttribute: 'string',
+          asArray: [],
+          asInt: 0,
+          asDecimal: 0,
+          asString: 'string',
+          asDate: '2021-12-14T14:10:31.030Z',
+          value: 'string'
+        },
+        children: []
+      }
+    ];
+
+    const getPrivate = true;
+
+    service.getDocumentPreviousQuestions(documentId, stationId, getPrivate)
+      .subscribe((responsePreviousQuestion) => {
+        expect(responsePreviousQuestion).toEqual(expectPreviousQuestions);
+      });
+
+    // eslint-disable-next-line max-len
+    const req = httpTestingController.expectOne(`${environment.baseAppUrl}${MICROSERVICE_PATH}/questions?documentRithmId=${documentId}&stationRithmId=${stationId}&getPrivate=${getPrivate}`);
+    expect(req.request.method).toEqual('GET');
+    expect(req.request.params.get('documentRithmId')).toBe(documentId);
+    expect(req.request.params.get('stationRithmId')).toBe(stationId);
+    expect(Boolean(req.request.params.get('getPrivate'))).toBeTrue();
+
+    req.flush(expectPreviousQuestions);
+    httpTestingController.verify();
+  });
+
+  it('should delete a document', () => {
+    service.deleteDocument(documentId)
+      .subscribe((response) => {
+        expect(response).toBeFalsy();
+      });
+
+    const req = httpTestingController.expectOne(`${environment.baseApiUrl}${MICROSERVICE_PATH}/${documentId}`);
+    expect(req.request.method).toEqual('DELETE');
+    req.flush(null);
+    httpTestingController.verify();
+  });
+
+  it('should return the user assigned to the document', () => {
+    const expectedResponse: StationRosterMember[] = [{
+      rithmId: '789-321-456-789',
+      firstName: 'John',
+      lastName: 'Christopher',
+      email: 'johnny.depp@gmail.com',
+      isAssigned: true
+    }];
+
+    service.getAssignedUserToDocument(documentId, stationId, true)
+      .subscribe((documentTimeInStation) => {
+        expect(documentTimeInStation).toEqual(expectedResponse);
+      });
+
+    // eslint-disable-next-line max-len
+    const req = httpTestingController.expectOne(`${environment.baseApiUrl}${MICROSERVICE_PATH}/assigned-user?documentId=${documentId}&stationId=${stationId}&getOnlyCurrentStation=true`);
+    expect(req.request.method).toEqual('GET');
+    expect(req.request.params.get('documentId')).toBe(documentId);
+    expect(req.request.params.get('stationId')).toBe(stationId);
+    expect(Boolean(req.request.params.get('getOnlyCurrentStation'))).toBeTrue();
+    req.flush(expectedResponse);
+    httpTestingController.verify();
+  });
+
+  it('should flow a document', () => {
+    const expectedData: DocumentAutoFlow = {
+      stationRithmId: stationId,
+      documentRithmId: documentId,
+      testMode: true
+    };
+
+    service.autoFlowDocument(expectedData)
+      .subscribe((response) => {
+        expect(response).toBeFalsy();
+      });
+
+    const req = httpTestingController.expectOne(`${environment.baseApiUrl}${MICROSERVICE_PATH}/auto-flow`);
+    expect(req.request.method).toEqual('POST');
+    expect(req.request.body).toEqual(expectedData);
+    req.flush(null);
+    httpTestingController.verify();
+  });
 });
