@@ -1,6 +1,6 @@
 import { HttpClient, HttpErrorResponse, HttpParams } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { BehaviorSubject, Observable, of, throwError } from 'rxjs';
+import { BehaviorSubject, Observable, Subject, throwError } from 'rxjs';
 import { delay, map } from 'rxjs/operators';
 import { environment } from 'src/environments/environment';
 // eslint-disable-next-line max-len
@@ -16,12 +16,20 @@ const MICROSERVICE_PATH = '/stationservice/api/station';
 })
 export class StationService {
 
-
   /** The Name of the Station as BehaviorSubject. */
   stationName$ = new BehaviorSubject<string>('');
 
+  /** Set the Question of the station-template which will be moved to previous fields expansion panel. */
+  questionToMove$ = new Subject<Question>();
+
   /** The Name of the Station Document as BehaviorSubject. */
   documentStationNameFields$ = new BehaviorSubject<DocumentNameField[]>([]);
+
+  /** Set touch to station template form. */
+  stationFormTouched$ = new Subject<void>();
+
+  /** The question to be updated when it changes in station page. */
+  stationQuestion$ = new Subject<Question>();
 
   constructor(
     private http: HttpClient
@@ -112,11 +120,12 @@ export class StationService {
   /**
    * Update the station questions.
    *
+   * @param stationId The Specific id of station.
    * @param questions The question to be updated.
    * @returns Station updated questions array.
    */
-  updateStationQuestions(questions: Question[]): Observable<Question[]> {
-    return this.http.post<Question[]>(`${environment.baseApiUrl}${MICROSERVICE_PATH}/questions`, questions);
+  updateStationQuestions(stationId: string, questions: Question[]): Observable<Question[]> {
+    return this.http.post<Question[]>(`${environment.baseApiUrl}${MICROSERVICE_PATH}/questions?stationRithmId=${stationId}`, questions);
   }
 
   /**
@@ -232,6 +241,15 @@ export class StationService {
   }
 
   /**
+   * Update the station question values in the template area.
+   *
+   * @param question The question to be updated.
+   */
+   updateStationQuestionInTemplate(question: Question): void {
+    this.stationQuestion$.next(question);
+  }
+
+  /**
    * Adds users to the owners roster.
    *
    * @param stationId The Specific id of station.
@@ -282,6 +300,15 @@ export class StationService {
     return this.http.get<boolean>(`${environment.baseApiUrl}${MICROSERVICE_PATH}/worker-rename-document`, { params });
   }
 
+ /**
+  * Reports a new question to be moved.
+  *
+  * @param question The question of the station-template to be moved.
+  */
+  moveQuestion(question: Question): void {
+    this.questionToMove$.next(question);
+  }
+
   /**
    * Update the Station General Instruction.
    *
@@ -299,40 +326,15 @@ export class StationService {
 
   /**
    * Update station name.
-   * Get previous and following stations.
+   * Get previous and next stations.
    *
    * @param stationRithmId The rithm id actually station.
-   * @returns Previous and following stations.
+   * @returns Previous and next stations.
    */
-  getPreviousAndFollowingStations(stationRithmId: string): Observable<ForwardPreviousStationsDocument> {
-    const data: ForwardPreviousStationsDocument = {
-      rithmId: stationRithmId,
-      previousStations: [
-        {
-          rithmId: '789-654-321',
-          name: 'Previous station 1',
-          totalDocuments: 5
-        },
-        {
-          rithmId: '789-654-753',
-          name: 'Previous station 2',
-          totalDocuments: 2
-        }
-      ],
-      followingStations: [
-        {
-          rithmId: '852-963-741',
-          name: 'Follow station 1',
-          totalDocuments: 2
-        },
-        {
-          rithmId: '852-963-418',
-          name: 'Follow station 2',
-          totalDocuments: 1
-        }
-      ]
-    };
-    return of(data).pipe(delay(1000));
+  getPreviousAndNextStations(stationRithmId: string): Observable<ForwardPreviousStationsDocument> {
+    const params = new HttpParams()
+      .set('stationRithmId', stationRithmId);
+    return this.http.get<ForwardPreviousStationsDocument>(`${environment.baseApiUrl}${MICROSERVICE_PATH}/prev-next-stations`, { params });
   }
 
   /**
@@ -373,5 +375,10 @@ export class StationService {
   updateDocumentNameTemplate(stationId: string, appendedFields: DocumentNameField[]): Observable<DocumentNameField[]> {
     // eslint-disable-next-line max-len
     return this.http.put<DocumentNameField[]>(`${environment.baseApiUrl}${MICROSERVICE_PATH}/document-naming-template?stationRithmId=${stationId}`, appendedFields);
+  }
+
+  /** Set touch to station template form. */
+  touchStationForm(): void {
+    this.stationFormTouched$.next();
   }
 }
