@@ -1,7 +1,7 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, fakeAsync, TestBed, tick } from '@angular/core/testing';
 import { MockComponent } from 'ng-mocks';
 import { ErrorService } from 'src/app/core/error.service';
-import { MockErrorService, MockStationService, MockUserService } from 'src/mocks';
+import { MockErrorService, MockStationService, MockUserService, MockDocumentService, MockPopupService } from 'src/mocks';
 import { DocumentInfoDrawerComponent } from './document-info-drawer.component';
 import { StationService } from 'src/app/core/station.service';
 import { MatCheckboxModule } from '@angular/material/checkbox';
@@ -13,12 +13,19 @@ import { MatInputModule } from '@angular/material/input';
 import { MatSelectModule } from '@angular/material/select';
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
 import { UserService } from 'src/app/core/user.service';
-
+import { DocumentService } from 'src/app/core/document.service';
+import { SidenavDrawerService } from 'src/app/core/sidenav-drawer.service';
+import { DialogOptions } from 'src/models';
+import { PopupService } from 'src/app/core/popup.service';
+import { RouterTestingModule } from '@angular/router/testing';
+import { throwError } from 'rxjs';
 
 describe('DocumentInfoDrawerComponent', () => {
   let component: DocumentInfoDrawerComponent;
   let fixture: ComponentFixture<DocumentInfoDrawerComponent>;
+  let sideNavService: SidenavDrawerService;
   const stationId = 'ED6148C9-ABB7-408E-A210-9242B2735B1C';
+  const documentId = 'E204F369-386F-4E41';
   const formBuilder = new FormBuilder();
 
   beforeEach(async () => {
@@ -31,7 +38,10 @@ describe('DocumentInfoDrawerComponent', () => {
         { provide: StationService, useClass: MockStationService },
         { provide: ErrorService, useClass: MockErrorService },
         { provide: FormGroup, useValue: formBuilder },
-        { provide: UserService, useClass: MockUserService }
+        { provide: UserService, useClass: MockUserService },
+        { provide: DocumentService, useClass: MockDocumentService },
+        { provide: SidenavDrawerService, useClass: SidenavDrawerService },
+        { provide: PopupService, useClass: MockPopupService }
       ],
       imports: [
         MatCheckboxModule,
@@ -41,7 +51,8 @@ describe('DocumentInfoDrawerComponent', () => {
         MatInputModule,
         NoopAnimationsModule,
         MatSelectModule,
-        FormsModule
+        FormsModule,
+        RouterTestingModule
       ],
     })
       .compileComponents();
@@ -50,6 +61,8 @@ describe('DocumentInfoDrawerComponent', () => {
   beforeEach(() => {
     fixture = TestBed.createComponent(DocumentInfoDrawerComponent);
     component = fixture.componentInstance;
+    sideNavService = TestBed.inject(SidenavDrawerService);
+    component.documentRithmId = documentId;
     fixture.detectChanges();
   });
 
@@ -73,5 +86,121 @@ describe('DocumentInfoDrawerComponent', () => {
     await component.getStatusDocumentEditable();
 
     expect(getGenerationStatusSpy).toHaveBeenCalledOnceWith(stationId);
+  });
+
+  it('should get document last updated date', () => {
+    const getLastUpdatedSpy = spyOn(TestBed.inject(DocumentService), 'getLastUpdated').and.callThrough();
+
+    sideNavService.drawerData$.next({
+      isStation: false,
+      documentRithmId: documentId
+    });
+
+    expect(getLastUpdatedSpy).toHaveBeenCalledOnceWith(documentId);
+  });
+
+  it('should get held time in station for document', () => {
+    const getDocumentTimeInStationSpy = spyOn(TestBed.inject(DocumentService), 'getDocumentTimeInStation').and.callThrough();
+
+    sideNavService.drawerData$.next({
+      isStation: false,
+      documentRithmId: documentId,
+      stationRithmId: stationId
+    });
+
+    expect(getDocumentTimeInStationSpy).toHaveBeenCalledOnceWith(documentId, stationId);
+  });
+
+  it('should return the user assigned to the document', () => {
+    const getAssignedUserSpy = spyOn(TestBed.inject(DocumentService), 'getAssignedUserToDocument').and.callThrough();
+    component.stationRithmId = stationId;
+    component['getAssignedUserToDocument']();
+
+    expect(getAssignedUserSpy).toHaveBeenCalledOnceWith(documentId, stationId, true);
+  });
+
+  it('should show loading-last-update while get data last updated', () => {
+    sideNavService.drawerData$.next({
+      isStation: false,
+      documentRithmId: documentId
+    });
+    fixture.detectChanges();
+    expect(component.lastUpdatedLoading).toBe(true);
+    const loadingComponent = fixture.debugElement.nativeElement.querySelector('#loading-last-update');
+    expect(loadingComponent).toBeTruthy();
+  });
+
+  it('should delete a document', async () => {
+    const deleteDocumentSpy = spyOn(TestBed.inject(DocumentService), 'deleteDocument').and.callThrough();
+
+    await component.deleteDocument();
+
+    expect(deleteDocumentSpy).toHaveBeenCalledOnceWith(documentId);
+  });
+
+
+  it('should open a confirm dialog to delete the document', async () => {
+    const dialogExpectData: DialogOptions = {
+      title: 'Are you sure?',
+      message: 'The document will be deleted.',
+      okButtonText: 'Delete',
+      cancelButtonText: 'Cancel',
+      important: true
+    };
+    const popupSpy = spyOn(TestBed.inject(PopupService), 'confirm').and.callThrough();
+
+    await component.deleteDocument();
+
+    expect(popupSpy).toHaveBeenCalledOnceWith(dialogExpectData);
+  });
+
+  it('should call the confirm dialog when clicking the delete button of document', fakeAsync(() => {
+    component.isUserAdminOrOwner = true;
+    fixture.detectChanges();
+    const deleteDocumentSpy = spyOn(component, 'deleteDocument');
+    const buttonDelete = fixture.debugElement.nativeElement.querySelector('button.priority');
+    expect(buttonDelete).toBeTruthy();
+    buttonDelete.click();
+    tick();
+    expect(deleteDocumentSpy).toHaveBeenCalledOnceWith();
+  }));
+
+  it('should show loading-indicators while get held time in station', () => {
+    sideNavService.drawerData$.next({
+      isStation: false,
+      documentRithmId: documentId,
+      stationRithmId: stationId
+    });
+    fixture.detectChanges();
+    expect(component.timeInStationLoading).toBeTrue();
+    const loadingComponent = fixture.debugElement.nativeElement.querySelector('#loading-time-in-station');
+    expect(loadingComponent).toBeTruthy();
+  });
+
+  it('should show assigned-user-loading while get assigned user of document', () => {
+    sideNavService.drawerData$.next({
+      isStation: false,
+      documentRithmId: documentId,
+      stationRithmId: stationId
+    });
+    fixture.detectChanges();
+    expect(component.assignedUserLoading).toBeTrue();
+    const loadingComponent = fixture.debugElement.nativeElement.querySelector('#assigned-user-loading');
+    expect(loadingComponent).toBeTruthy();
+  });
+
+  it('should show error message when request for assigned user fails', () => {
+    spyOn(TestBed.inject(DocumentService), 'getAssignedUserToDocument').and.returnValue(throwError(() => {
+      throw new Error();
+    }));
+    sideNavService.drawerData$.next({
+      isStation: false,
+      documentRithmId: '',
+      stationRithmId: stationId
+    });
+    fixture.detectChanges();
+    expect(component.userErrorAssigned).toBeTrue();
+    const errorComponent = fixture.debugElement.nativeElement.querySelector('#assigned-user-error');
+    expect(errorComponent).toBeTruthy();
   });
 });
