@@ -8,9 +8,10 @@ import { UtcTimeConversion } from 'src/helpers';
 import { ActivatedRoute, Router } from '@angular/router';
 import { SidenavDrawerService } from 'src/app/core/sidenav-drawer.service';
 import { UserService } from 'src/app/core/user.service';
-import { DocumentGenerationStatus, StationInfoDrawerData, StationInformation } from 'src/models';
+import { DocumentGenerationStatus, MapItemStatus, MapMode, StationInfoDrawerData, StationInformation } from 'src/models';
 import { PopupService } from 'src/app/core/popup.service';
 import { MatRadioChange } from '@angular/material/radio';
+import { MapService } from 'src/app/map/map.service';
 
 /**
  * Component for info station.
@@ -44,11 +45,23 @@ export class StationInfoDrawerComponent implements OnInit, OnDestroy {
   /** Is component viewed in station edit mode. */
   editMode = false;
 
-  /** Station information object passed from parent. */
+  /** Station information object. */
   stationInformation!: StationInformation;
+
+  /** Station Id passed from parent. */
+  stationRithmId = '';
 
   /** Edit Mode. */
   stationName = '';
+
+  /** Notes for the station. */
+  stationNotes?: string;
+
+  /** If component is being viewed on the map, what mode is the map in? */
+  mapMode?: MapMode;
+
+  /** If component is being viewed on the map, what status does the station have? */
+  stationStatus?: MapItemStatus;
 
   /** Station name form. */
   stationNameForm: FormGroup;
@@ -62,6 +75,15 @@ export class StationInfoDrawerComponent implements OnInit, OnDestroy {
   /** Color message LastUpdated. */
   colorMessage = '';
 
+  /** Whether the station drawer is opened from map or not. */
+  openedFromMap = false;
+
+  /** Allowing access to all MapMode enums in HTML.*/
+  mapModeEnum = MapMode;
+
+  /** The priority for current station once the info is loaded.*/
+  stationPriority: number | '--' = '--';
+
   constructor(
     private sidenavDrawerService: SidenavDrawerService,
     private userService: UserService,
@@ -71,17 +93,28 @@ export class StationInfoDrawerComponent implements OnInit, OnDestroy {
     private errorService: ErrorService,
     private route: ActivatedRoute,
     private popupService: PopupService,
-    private router: Router
+    private router: Router,
+    private mapService: MapService
   ) {
     this.sidenavDrawerService.drawerData$
       .pipe(takeUntil(this.destroyed$))
-      .subscribe((data) => {
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any
+      .subscribe((data: any) => {
         const dataDrawer = data as StationInfoDrawerData;
         if (dataDrawer) {
           this.editMode = dataDrawer.editMode;
-          this.stationInformation = dataDrawer.stationInformation as StationInformation;
+          this.stationRithmId = dataDrawer.stationRithmId;
           this.stationName = dataDrawer.stationName;
+          this.mapMode = dataDrawer.mapMode;
+          this.stationStatus = dataDrawer.stationStatus;
+          this.openedFromMap = dataDrawer.openedFromMap;
+          this.stationNotes = dataDrawer.notes;
+          this.editMode = dataDrawer.editMode;
+          if (this.openedFromMap && this.stationStatus !== MapItemStatus.Created) {
+            this.getStationDocumentGenerationStatus();
+          }
         }
+        this.getStationInfo();
       });
 
     this.type = this.userService.user.role === 'admin' ? this.userService.user.role : 'worker';
@@ -94,31 +127,40 @@ export class StationInfoDrawerComponent implements OnInit, OnDestroy {
    * Gets info about the station as well as forward and previous stations for a specific station.
    */
   ngOnInit(): void {
-    this.getParams();
-    this.getStationDocumentGenerationStatus(this.stationInformation.rithmId);
+    if (this.stationStatus !== MapItemStatus.Created) {
+      this.getLastUpdated();
+      this.getStationDocumentGenerationStatus();
 
-    this.stationService.stationName$
-      .pipe(takeUntil(this.destroyed$))
-      .subscribe({
-        next: (data) => {
-          this.stationName = data.length > 0 ? data : 'Untitled Station';
-        }, error: (error: unknown) => {
-          this.errorService.displayError(
-            'Something went wrong on our end and we\'re looking into it. Please try again in a little while.',
-            error
-          );
-        }
-      });
+      this.stationService.stationName$
+        .pipe(takeUntil(this.destroyed$))
+        .subscribe({
+          next: (data) => {
+            this.stationName = data.length > 0 ? data : 'Untitled Station';
+          }, error: (error: unknown) => {
+            this.errorService.displayError(
+              'Something went wrong on our end and we\'re looking into it. Please try again in a little while.',
+              error
+            );
+          }
+        });
+    }
+  }
+
+  /**
+   * Whether the station is locally created on the map.
+   *
+   * @returns True if locally created, false otherwise.
+   */
+  get locallyCreated(): boolean {
+    return this.stationStatus === MapItemStatus.Created;
   }
 
   /**
    * Get station document generation status.
-   *
-   * @param stationId The id of the station return status document.
    */
-  getStationDocumentGenerationStatus(stationId: string): void {
+  getStationDocumentGenerationStatus(): void {
     this.docGenLoading = true;
-    this.stationService.getStationDocumentGenerationStatus(stationId)
+    this.stationService.getStationDocumentGenerationStatus(this.stationRithmId)
       .pipe(first())
       .subscribe({
         next: (status: DocumentGenerationStatus) => {
@@ -164,36 +206,12 @@ export class StationInfoDrawerComponent implements OnInit, OnDestroy {
   }
 
   /**
-   * Attempts to retrieve the station info from the query params in the URL and make the requests.
-   */
-  private getParams(): void {
-    this.route.params
-      .pipe(first())
-      .subscribe({
-        next: (params) => {
-          if (!params.stationId) {
-            this.handleInvalidParams();
-          } else {
-            this.getLastUpdated(params.stationId);
-          }
-        }, error: (error: unknown) => {
-          this.errorService.displayError(
-            'Something went wrong on our end and we\'re looking into it. Please try again in a little while.',
-            error
-          );
-        }
-      });
-  }
-
-  /**
    * Get the last updated date for a specific station.
-   *
-   * @param stationId The id of the station that the document is in.
    */
-  getLastUpdated(stationId: string): void {
+  getLastUpdated(): void {
     this.stationLoading = true;
     this.lastUpdatedLoading = true;
-    this.stationService.getLastUpdated(stationId)
+    this.stationService.getLastUpdated(this.stationRithmId)
       .pipe(first())
       .subscribe({
         next: (updatedDate) => {
@@ -227,22 +245,10 @@ export class StationInfoDrawerComponent implements OnInit, OnDestroy {
   }
 
   /**
-   * Navigates the user back to dashboard and displays a message about the invalid params.
-   */
-  private handleInvalidParams(): void {
-    this.errorService.displayError(
-      'Unable to retrieve the last updated time.',
-      new Error('Invalid params for document')
-    );
-  }
-
-  /**
    * This will delete the current station.
-   *
-   * @param stationId Target station to be deleted.
    */
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  async deleteStation(stationId: string): Promise<void> {
+  async deleteStation(): Promise<void> {
     const response = await this.popupService.confirm({
       title: 'Are you sure?',
       message: 'The station will be deleted for everyone and any documents not moved to another station beforehand will be deleted.',
@@ -251,7 +257,7 @@ export class StationInfoDrawerComponent implements OnInit, OnDestroy {
       important: true,
     });
     if (response) {
-      this.stationService.deleteStation(stationId)
+      this.stationService.deleteStation(this.stationRithmId)
         .pipe(first())
         .subscribe({
           next: () => {
@@ -268,20 +274,12 @@ export class StationInfoDrawerComponent implements OnInit, OnDestroy {
   }
 
   /**
-   * Completes all subscriptions.
-   */
-  ngOnDestroy(): void {
-    this.destroyed$.next();
-    this.destroyed$.complete();
-  }
-
-  /**
    * Update status the station.
    *
    * @param statusNew New status the station update.
    */
   updateStatusStation(statusNew: MatRadioChange): void {
-    this.updateStationDocumentGenerationStatus(this.stationInformation.rithmId, statusNew.value);
+    this.updateStationDocumentGenerationStatus(this.stationRithmId, statusNew.value);
   }
 
   /**
@@ -290,23 +288,29 @@ export class StationInfoDrawerComponent implements OnInit, OnDestroy {
    */
   getStationInfo(): void {
     this.stationLoading = true;
-    this.stationService.getStationInfo(this.stationInformation.rithmId)
-      .pipe(first())
-      .subscribe({
-        next: (stationInfo) => {
-          this.stationLoading = false;
-          if (stationInfo) {
-            this.stationInformation = stationInfo;
+    if (this.stationStatus !== MapItemStatus.Created) {
+      this.stationService.getStationInfo(this.stationRithmId)
+        .pipe(first())
+        .subscribe({
+          next: (stationInfo) => {
+            this.stationLoading = false;
+            if (stationInfo) {
+              this.stationInformation = stationInfo;
+              this.stationPriority = stationInfo.priority;
+            }
+          },
+          error: (error: unknown) => {
+            this.stationLoading = false;
+            this.errorService.displayError(
+              'Something went wrong on our end and we\'re looking into it. Please try again in a little while.',
+              error
+            );
           }
-        },
-        error: (error: unknown) => {
-          this.stationLoading = false;
-          this.errorService.displayError(
-            'Something went wrong on our end and we\'re looking into it. Please try again in a little while.',
-            error
-          );
-        }
-      });
+        });
+    } else {
+      this.stationLoading = false;
+      //Is there any information for a new station that needs to be populated here?
+    }
   }
 
   /**
@@ -319,4 +323,58 @@ export class StationInfoDrawerComponent implements OnInit, OnDestroy {
       this.getStationInfo();
     }
   }
+
+  /**
+   * Handle required information for a locally created station.
+   */
+  newStationInit(): void {
+    this.stationDocumentGenerationStatus = DocumentGenerationStatus.None;
+    this.lastUpdatedDate = 'Publish Map changes to see last updated.';
+  }
+
+  /**
+   * Navigate to station edit page upon confirmation in Map build mode and without any confirmation in Map view mode.
+   *
+   */
+  async goToStation(): Promise<void> {
+    let confirmNavigation = false;
+    if (this.editMode) {
+      const confirm = await this.popupService.confirm({
+        title: 'Local Changes Not Saved',
+        message: `Leave without publishing any changes made to the map?`,
+        okButtonText: 'Proceed',
+      });
+      confirmNavigation = confirm;
+    }
+    if (confirmNavigation || !this.editMode) {
+      this.router.navigate([`/station/${this.stationRithmId}`]);
+    }
+  }
+
+  /**
+   * Reporting if the name or notes on a station changed.
+   */
+  reportNewStationMapChange(): void {
+    if (this.stationNotes === undefined) {
+      throw new Error('Station notes not found');
+    }
+    const openStation = this.mapService.stationElements.find((station) => this.stationInformation.rithmId === station.rithmId);
+    if (openStation === undefined) {
+      throw new Error('Station was not found.');
+    }
+    this.stationName = this.stationNameForm.value.name;
+    openStation.stationName = this.stationName;
+    openStation.notes = this.stationNotes;
+    openStation.markAsUpdated();
+    this.mapService.stationElementsChanged$.next(true);
+  }
+
+  /**
+   * Completes all subscriptions.
+   */
+   ngOnDestroy(): void {
+    this.destroyed$.next();
+    this.destroyed$.complete();
+  }
+
 }
