@@ -5,9 +5,11 @@ import { StationMapElement } from 'src/helpers';
 import { MapMode, Point, MapDragItem, MapItemStatus, FlowMapElement, StationElementHoverType,
   StationInfoDrawerData, StationInformation, ConnectionMapElement } from 'src/models';
 import { ConnectionElementService } from '../connection-element.service';
-import { DEFAULT_MOUSE_POINT, DEFAULT_SCALE, MAX_SCALE, MIN_SCALE,
+import {
+  DEFAULT_MOUSE_POINT, DEFAULT_SCALE, MAX_SCALE, MIN_SCALE,
   PAN_DECAY_RATE, PAN_TRIGGER_LIMIT, SCALE_RENDER_STATION_ELEMENTS,
-  STATION_HEIGHT, STATION_WIDTH, ZOOM_VELOCITY, MAX_PAN_VELOCITY } from '../map-constants';
+  STATION_HEIGHT, STATION_WIDTH, ZOOM_VELOCITY, MAX_PAN_VELOCITY, MOUSE_MOVEMENT_OVER_CONNECTION
+} from '../map-constants';
 import { MapService } from '../map.service';
 import { StationElementService } from '../station-element.service';
 import { FlowElementService } from '../flow-element.service';
@@ -15,7 +17,6 @@ import { StationDocumentsModalComponent } from 'src/app/shared/station-documents
 import { MatDialog } from '@angular/material/dialog';
 import { SidenavDrawerService } from 'src/app/core/sidenav-drawer.service';
 import { StationService } from 'src/app/core/station.service';
-import { PopupService } from 'src/app/core/popup.service';
 
 /**
  * Component for the main `<canvas>` element used for the map.
@@ -83,6 +84,9 @@ export class MapCanvasComponent implements OnInit, OnDestroy {
   /** Data for connection line path between stations. */
   connections: ConnectionMapElement[] = [];
 
+  /** Initial Data Load. */
+  private initLoad = true;
+
   /** Scale to calculate canvas points. */
   private scale = DEFAULT_SCALE;
 
@@ -91,6 +95,9 @@ export class MapCanvasComponent implements OnInit, OnDestroy {
 
   /**Set up interval for zoom. */
   private zoomInterval?: NodeJS.Timeout;
+
+  /** Boolean to check drag on connection line. */
+  private connectionLineDrag = false;
 
 
 
@@ -110,8 +117,7 @@ export class MapCanvasComponent implements OnInit, OnDestroy {
     private flowElementService: FlowElementService,
     private dialog: MatDialog,
     private sidenavDrawerService: SidenavDrawerService,
-    private stationService: StationService,
-    private popupService: PopupService
+    private stationService: StationService
   ) {
     this.mapService.mapMode$
       .pipe(takeUntil(this.destroyed$))
@@ -134,12 +140,10 @@ export class MapCanvasComponent implements OnInit, OnDestroy {
         this.drawElements();
       });
 
-    this.mapService.mapDataReceived$
+    this.mapService.stationElementsChanged$
       .pipe(takeUntil(this.destroyed$))
       .subscribe(() => {
-        this.stations = this.mapService.stationElements.filter((e) => e.status !== MapItemStatus.Deleted);
-        this.flows = this.mapService.flowElements;
-        this.connections = this.mapService.connectionElements;
+        this.stations = this.mapService.stationElements;
         this.drawElements();
       });
 
@@ -169,6 +173,18 @@ export class MapCanvasComponent implements OnInit, OnDestroy {
     this.context = this.mapCanvas.nativeElement.getContext('2d') as CanvasRenderingContext2D;
     this.mapService.registerCanvasContext(this.context);
     this.setCanvasSize();
+    this.mapService.mapDataReceived$
+    .pipe(takeUntil(this.destroyed$))
+    .subscribe((dataReceived) => {
+      this.stations = this.mapService.stationElements.filter((e) => e.status !== MapItemStatus.Deleted);
+      this.flows = this.mapService.flowElements;
+      this.connections = this.mapService.connectionElements;
+      if (dataReceived && this.initLoad) {
+        this.mapService.center(dataReceived);
+        this.initLoad = false;
+      }
+      this.drawElements();
+    });
     this.drawElements();
   }
 
@@ -220,9 +236,7 @@ export class MapCanvasComponent implements OnInit, OnDestroy {
           const pointer = this.pointerCache[0];
           this.lastTouchCoords[0] = this.getEventCanvasPoint(pointer);
           this.eventStartCoords = this.getEventCanvasPoint(pointer);
-
-          const pos = this.getEventCanvasPoint(pointer);
-          this.eventStartLogic(pos);
+          this.eventStartLogic(event);
         }
 
         if (this.pointerCache.length === 2) {
@@ -256,7 +270,7 @@ export class MapCanvasComponent implements OnInit, OnDestroy {
     if (window.PointerEvent && !(is_android && is_firefox)) {
       event.preventDefault();
 
-      //remove event from cache.
+      // remove event from cache.
       for (let i = 0; i < this.pointerCache.length; i++) {
         if (this.pointerCache[i].pointerId === event.pointerId) {
           this.pointerCache.splice(i, 1);
@@ -265,15 +279,12 @@ export class MapCanvasComponent implements OnInit, OnDestroy {
       }
 
       if (this.pointerCache.length === 0) {
-        const pos = this.getEventCanvasPoint(event);
-        const con = this.getEventContextPoint(event);
-
         if (this.dragItem !== MapDragItem.Default) {
           const map = document.getElementById('map');
           map?.releasePointerCapture(event.pointerId);
         }
 
-        this.eventEndLogic(pos, con);
+        this.eventEndLogic(event);
       } else {
         const pointer = this.pointerCache[0];
         this.lastTouchCoords[0] = this.getEventCanvasPoint(pointer);
@@ -340,9 +351,7 @@ export class MapCanvasComponent implements OnInit, OnDestroy {
       this.eventStartCoords = this.getEventCanvasPoint(event);
       this.lastTouchCoords[0] = this.getEventCanvasPoint(event);
       this.eventStartCoords = this.getEventCanvasPoint(event);
-
-      const pos = this.getEventCanvasPoint(event);
-      this.eventStartLogic(pos);
+      this.eventStartLogic(event);
     }
   }
 
@@ -355,11 +364,8 @@ export class MapCanvasComponent implements OnInit, OnDestroy {
   @HostListener('mouseup', ['$event'])
   mouseUp(event: MouseEvent): void {
     if (!window.PointerEvent) {
-      const pos = this.getEventCanvasPoint(event);
       this.lastTouchCoords[0] = this.getEventCanvasPoint(event);
-      const con = this.getEventContextPoint(event);
-
-      this.eventEndLogic(pos, con);
+      this.eventEndLogic(event);
     }
   }
 
@@ -396,12 +402,12 @@ export class MapCanvasComponent implements OnInit, OnDestroy {
     if (!window.PointerEvent || (is_android && is_firefox)) {
       if (event.touches.length === 1) {
         const touchPoint = event.touches[0];
-        const pos = this.getEventCanvasPoint(touchPoint);
+        const eventCanvasPoint = this.getEventCanvasPoint(touchPoint);
 
-        this.lastTouchCoords[0] = pos;
-        this.eventStartCoords = pos;
+        this.lastTouchCoords[0] = eventCanvasPoint;
+        this.eventStartCoords = eventCanvasPoint;
 
-        this.eventStartLogic(pos);
+        this.eventStartLogic(touchPoint);
       }
 
       if (event.touches.length === 2) {
@@ -432,10 +438,7 @@ export class MapCanvasComponent implements OnInit, OnDestroy {
 
     if (!window.PointerEvent || (is_android && is_firefox)) {
       const touchPoint = event.changedTouches[0];
-      const pos = this.getEventCanvasPoint(touchPoint);
-      const con = this.getEventContextPoint(touchPoint);
-
-      this.eventEndLogic(pos, con);
+      this.eventEndLogic(touchPoint);
     }
   }
 
@@ -505,7 +508,7 @@ export class MapCanvasComponent implements OnInit, OnDestroy {
     if (event.key === '+' || event.key === '=' || event.key === '-') {
       this.mapService.matMenuStatus$.next(true);
       this.mapService.zoomCount$.next(this.zoomCount + (event.key === '+' || event.key === '=' ? 50 : -50));
-      this.mapService.handleZoom(undefined, false);
+      this.mapService.handleZoom(false);
     }
   }
 
@@ -535,7 +538,7 @@ export class MapCanvasComponent implements OnInit, OnDestroy {
         this.mapService.zoomCount$.next(0);
       }
       this.mapService.zoomCount$.next(this.zoomCount + Math.floor(10*-eventAmount));
-      this.mapService.handleZoom(mousePoint, false);
+      this.mapService.handleZoom(false, mousePoint);
     } else {
       // Do nothing if already at min zoom.
       if (this.scale <= MIN_SCALE
@@ -549,7 +552,7 @@ export class MapCanvasComponent implements OnInit, OnDestroy {
         this.mapService.zoomCount$.next(0);
       }
       this.mapService.zoomCount$.next(this.zoomCount - Math.floor(10*eventAmount));
-      this.mapService.handleZoom(mousePoint, false);
+      this.mapService.handleZoom(false, mousePoint);
     }
     // Overlay option menu close state.
     if (this.mapService.matMenuStatus$ && this.mapMode === MapMode.Build) {
@@ -756,9 +759,11 @@ export class MapCanvasComponent implements OnInit, OnDestroy {
   /**
    * Handles mouseDown and touchStart logic.
    *
-   * @param position The position of the mouse or touch event.
+   * @param event Is an input event.
    */
-  private eventStartLogic(position: Point) {
+  private eventStartLogic(event: MouseEvent | Touch) {
+    const eventCanvasPoint = this.getEventCanvasPoint(event);
+    const eventContextPoint = this.getEventContextPoint(event);
     if (this.panActive) {
       cancelAnimationFrame(this.myReq as number);
       this.panActive = false;
@@ -773,7 +778,7 @@ export class MapCanvasComponent implements OnInit, OnDestroy {
     if (this.mapMode === MapMode.Build) {
       for (const station of this.stations) {
         // Check if clicked on an interactive station element.
-        station.checkElementHover(position, this.mapMode, this.scale);
+        station.checkElementHover(eventCanvasPoint, this.mapMode, this.scale);
         // clicked on a connection node.
         if (station.hoverActive === StationElementHoverType.Node) {
           station.dragging = true;
@@ -789,6 +794,20 @@ export class MapCanvasComponent implements OnInit, OnDestroy {
         }
       }
 
+      for (const connection of this.connections) {
+        // Check if connection line was clicked. ContextPoint is used for connection lines.
+        connection.checkElementHover(eventContextPoint, this.context);
+        if (connection.hoverActive) {
+          const startStation = this.stations.find((station) => station.rithmId === connection.startStationRithmId);
+          if (!startStation) {
+            throw new Error(`Unable to find a start station with the id of ${connection.startStationRithmId} for a connection`);
+          }
+          startStation.dragging = true;
+         this.dragItem = MapDragItem.Connection;
+         break;
+        }
+      }
+
       //This ensures that when dragging a station or node connection, it will always display above other stations.
       if (this.stations.find( obj => obj.dragging === true)) {
         const draggingStation = this.stations.filter( obj => obj.dragging === true);
@@ -797,7 +816,7 @@ export class MapCanvasComponent implements OnInit, OnDestroy {
       }
     }
 
-    if (this.dragItem !== MapDragItem.Station && this.dragItem !== MapDragItem.Node) {
+    if (this.dragItem === MapDragItem.Default) {
       // Assume map for now
       this.dragItem = MapDragItem.Map;
     }
@@ -806,10 +825,11 @@ export class MapCanvasComponent implements OnInit, OnDestroy {
   /**
    * Handles mouseUp and touchEnd logic.
    *
-   * @param position The position of the mouse or touch event.
-   * @param contextPoint Calculated position of click.
+   * @param event Is a mouse or touch event.
    */
-  private eventEndLogic(position: Point, contextPoint: Point) {
+  private eventEndLogic(event: MouseEvent | Touch) {
+    const eventCanvasPoint = this.getEventCanvasPoint(event);
+    const eventContextPoint = this.getEventContextPoint(event);
     this.holdDrag = false;
 
     // Overlay option menu close state.
@@ -818,10 +838,15 @@ export class MapCanvasComponent implements OnInit, OnDestroy {
     }
 
     //If it is a click and not a drag.
-    if (Math.abs(position.x - this.eventStartCoords.x) < 5 && Math.abs(position.y - this.eventStartCoords.y) < 5) {
+    if (Math.abs(eventCanvasPoint.x - this.eventStartCoords.x) < 5 && Math.abs(eventCanvasPoint.y - this.eventStartCoords.y) < 5) {
+      this.dragItem = MapDragItem.Default;
+      this.stations.forEach((station) => {
+        station.dragging = false;
+      });
       if (this.scale >= SCALE_RENDER_STATION_ELEMENTS) {
-        this.clickEventHandler(position, contextPoint);
+        this.clickEventHandler(eventCanvasPoint, eventContextPoint);
       }
+      return;
     }
 
     //If dragging the map.
@@ -841,7 +866,7 @@ export class MapCanvasComponent implements OnInit, OnDestroy {
       let newPreviousStation: StationMapElement | undefined;
       for (const station of this.stations) {
         // Check if clicked on an interactive station element.
-        station.checkElementHover(position, this.mapMode, this.scale);
+        station.checkElementHover(eventCanvasPoint, this.mapMode, this.scale);
         if (station.hoverActive !== StationElementHoverType.None) {
           newNextStation = station;
         }
@@ -853,7 +878,7 @@ export class MapCanvasComponent implements OnInit, OnDestroy {
       if (newNextStation && newPreviousStation) {
         for (const station of this.stations) {
           // Check if clicked on an interactive station element.
-          station.checkElementHover(position, this.mapMode, this.scale);
+          station.checkElementHover(eventCanvasPoint, this.mapMode, this.scale);
           if (station.hoverActive === StationElementHoverType.Station) {
             //ensure we cant get duplicate ids.
             if (!station.previousStations.includes(newPreviousStation.rithmId) && station.rithmId !== newPreviousStation.rithmId) {
@@ -898,6 +923,7 @@ export class MapCanvasComponent implements OnInit, OnDestroy {
     this.eventStartCoords = DEFAULT_MOUSE_POINT;
     this.lastTouchCoords = [DEFAULT_MOUSE_POINT];
     this.mapCanvas.nativeElement.style.cursor = 'default';
+    this.connectionLineDrag = false;
   }
 
   /**
@@ -947,6 +973,22 @@ export class MapCanvasComponent implements OnInit, OnDestroy {
         // Check if clicked on an interactive station element.
         station.checkElementHover(this.mapService.currentMousePoint$.value, this.mapMode, this.scale);
         if (station.dragging) {
+          this.mapService.currentMousePoint$.next(moveInput);
+        }
+      }
+      // Check for Add New Connected Station mode is enabled or not also draw a temporary line from station's node.
+    } else if (this.mapMode === MapMode.StationAdd && this.mapService.stationElements.some(e => e.isAddingConnected)) {
+        this.mapService.currentMousePoint$.next(moveInput);
+    } else if (this.dragItem === MapDragItem.Connection) {
+      //If it is a drag and not a click.
+      const moveFromStartX = this.eventStartCoords.x - moveInput.x;
+      const moveFromStartY = this.eventStartCoords.y - moveInput.y;
+      if (Math.abs(moveFromStartX) > MOUSE_MOVEMENT_OVER_CONNECTION || Math.abs(moveFromStartY) > MOUSE_MOVEMENT_OVER_CONNECTION) {
+        this.onConnectionDrag(moveContext);
+      }
+      if (this.connectionLineDrag) {
+        this.mapCanvas.nativeElement.style.cursor = 'grabbing';
+        if (this.stations.some((station) => station.dragging)) {
           this.mapService.currentMousePoint$.next(moveInput);
         }
       }
@@ -1019,12 +1061,12 @@ export class MapCanvasComponent implements OnInit, OnDestroy {
       // Zoom in
       this.lastTouchCoords = position;
       this.mapService.zoomCount$.next(this.zoomCount + averageDiff);
-      this.mapService.handleZoom(middlePoint, true);
+      this.mapService.handleZoom(true, middlePoint);
     } else if (averageEnd < averageStart) {
       // Zoom out
       this.lastTouchCoords = position;
       this.mapService.zoomCount$.next(this.zoomCount + averageDiff);
-      this.mapService.handleZoom(middlePoint, true);
+      this.mapService.handleZoom(true, middlePoint);
     }
   }
 
@@ -1095,46 +1137,72 @@ export class MapCanvasComponent implements OnInit, OnDestroy {
   }
 
   /**
+   * Handles when a user drags an existing connection line.
+   *
+   * @param contextPoint Calculated position of click.
+   */
+  onConnectionDrag(contextPoint: Point): void {
+     for (const connectionLine of this.connections) {
+       connectionLine.checkElementHover(contextPoint, this.context);
+       if (connectionLine.hoverActive && !this.connectionLineDrag) {
+         // Created for future tasks.
+         // const startStation = this.stations.find(station => station.rithmId === connectionLine.startStationRithmId);
+         // const endStation = this.stations.find(station => station.rithmId === connectionLine.endStationRithmId);
+         // if ( !startStation || !endStation ){
+         //   throw new Error('This start or end station was not found.');
+         // }
+         // const storedConnectionLine = new ConnectionMapElement(startStation, endStation, this.scale);
+         this.mapService.removeConnectionLine(connectionLine.startStationRithmId, connectionLine.endStationRithmId);
+         this.connectionLineDrag = true;
+         break;
+       }
+     }
+  }
+
+  /**
    * Handles user input on a clicked connection line.
    *
    * @param station The clicked station.
    */
   checkStationClick(station: StationMapElement): void {
     // TODO: Remove this test rename prompt once renaming in the drawer is done
-            // this.popupService.prompt({
-            //   title: 'Rename Station',
-            //   message: 'Please provide a name for this station',
-            //   promptLabel: 'Station name',
-            //   promptValue: station.stationName
-            // }).then((newName) => {
-            //   if (newName && newName !== station.stationName) {
-            //     station.stationName = newName;
-            //     station.markAsUpdated();
-            //     this.drawElements();
-            //   }
-            // });
-            const stationDataInfo: StationInformation = {
-              rithmId: station.rithmId,
-              name: '',
-              instructions: '',
-              nextStations: [],
-              previousStations: [],
-              stationOwners: [],
-              workers: [],
-              createdByRithmId: '',
-              createdDate: '',
-              updatedByRithmId: '',
-              updatedDate: '',
-              questions: [],
-              priority: 1
-            };
-          const dataInformationDrawer: StationInfoDrawerData = {
-            stationInformation: stationDataInfo,
-            stationName: station.stationName,
-            editMode: this.mapMode === MapMode.Build,
-            locallyCreated: station.status === MapItemStatus.Created
-          };
-          this.sidenavDrawerService.openDrawer('stationInfo', dataInformationDrawer);
-          this.stationService.updatedStationNameText(station.stationName);
+    // this.popupService.prompt({
+    //   title: 'Rename Station',
+    //   message: 'Please provide a name for this station',
+    //   promptLabel: 'Station name',
+    //   promptValue: station.stationName
+    // }).then((newName) => {
+    //   if (newName && newName !== station.stationName) {
+    //     station.stationName = newName;
+    //     station.markAsUpdated();
+    //     this.drawElements();
+    //   }
+    // });
+    const stationDataInfo: StationInformation = {
+      rithmId: station.rithmId,
+      name: '',
+      instructions: '',
+      nextStations: [],
+      previousStations: [],
+      stationOwners: [],
+      workers: [],
+      createdByRithmId: '',
+      createdDate: '',
+      updatedByRithmId: '',
+      updatedDate: '',
+      questions: [],
+      priority: 1
+    };
+    const dataInformationDrawer: StationInfoDrawerData = {
+      stationRithmId: stationDataInfo.rithmId,
+      stationName: station.stationName,
+      editMode: this.mapMode === MapMode.Build,
+      stationStatus: station.status,
+      mapMode: this.mapMode,
+      openedFromMap: true,
+      notes: station.notes,
+    };
+    this.sidenavDrawerService.openDrawer('stationInfo', dataInformationDrawer);
+    this.stationService.updatedStationNameText(station.stationName);
   }
 }
