@@ -4,7 +4,7 @@ import { HttpClient } from '@angular/common/http';
 import { MapMode, Point, MapData, MapItemStatus, FlowMapElement, EnvironmentName, ConnectionMapElement } from 'src/models';
 import { ABOVE_MAX, BELOW_MIN, DEFAULT_CANVAS_POINT, DEFAULT_SCALE,
   MAX_SCALE, MIN_SCALE, SCALE_RENDER_STATION_ELEMENTS,
-  ZOOM_VELOCITY, DEFAULT_MOUSE_POINT } from './map-constants';
+  ZOOM_VELOCITY, DEFAULT_MOUSE_POINT, STATION_WIDTH, STATION_HEIGHT, SCALE_REDUCED_RENDER, CENTER_ZOOM_BUFFER } from './map-constants';
 import { environment } from 'src/environments/environment';
 import { map } from 'rxjs/operators';
 import { v4 as uuidv4 } from 'uuid';
@@ -26,6 +26,9 @@ export class MapService {
 
   /** Notifies when the map data has been received. */
   mapDataReceived$ = new BehaviorSubject(false);
+
+  /** Informs the map when station elements have changed. */
+  stationElementsChanged$ = new BehaviorSubject(false);
 
   /** The station elements displayed on the map. */
   stationElements: StationMapElement[] = [];
@@ -81,11 +84,11 @@ export class MapService {
   }
 
   /**
-   * Gets all map elements for a given organization.
+   * Gets all data for the map for a given organization.
    *
-   * @returns Retrieves all map elements for a given organization.
+   * @returns The map data for the organization.
    */
-  getMapElements(): Observable<MapData> {
+  getMapData(): Observable<MapData> {
     return this.http.get<MapData>(`${environment.baseApiUrl}${MICROSERVICE_PATH}/all`)
       .pipe(map((data) => {
         data.stations.map((e) => {
@@ -129,12 +132,14 @@ export class MapService {
    * Fills in connections array with info from this.stationElements.
    */
   setConnections(): void {
+    this.connectionElements = [];
     for (const station of this.stationElements) {
       for (const connection of station.nextStations) {
         const outgoingStation = this.stationElements.find((foundStation) => foundStation.rithmId === connection);
 
         if (!outgoingStation) {
-          throw new Error('no outgoing station found.');
+          throw new Error(`An outgoing station was not found for the stationId: ${connection} which appears in the
+            nextStations of the station${station.stationName}: ${station.rithmId}.`);
         }
 
         const lineInfo = new ConnectionMapElement(station, outgoingStation, this.mapScale$.value);
@@ -178,7 +183,33 @@ export class MapService {
       previousStations: [],
       nextStations: [],
       status: MapItemStatus.Created,
+      notes: '',
     });
+
+    // Connected station create changes
+    const connectedStations = this.stationElements.filter(station => station.isAddingConnected);
+    if (connectedStations.length === 1) {
+      const stationIndex = this.stationElements.findIndex(station => station.rithmId === connectedStations[0].rithmId);
+      const flowIndex = this.flowElements.findIndex(flow => flow.stations.includes(connectedStations[0].rithmId));
+      if (stationIndex >= 0) {
+        this.stationElements[stationIndex].isAddingConnected = false;
+        this.stationElements[stationIndex].nextStations.push(newStation.rithmId);
+        newStation.previousStations.push(this.stationElements[stationIndex].rithmId);
+
+        const lineInfo = new ConnectionMapElement(this.stationElements[stationIndex], newStation, this.mapScale$.value);
+        if (!this.connectionElements.includes(lineInfo)) {
+          this.connectionElements.push(lineInfo);
+        }
+        this.mapMode$.next(MapMode.Build);
+        this.stationElements[stationIndex].markAsUpdated();
+
+        if (flowIndex >= 0 && (!this.flowElements[flowIndex].stations.includes(newStation.rithmId))) {
+          this.flowElements[flowIndex].stations.push(newStation.rithmId);
+          this.flowElements[flowIndex].markAsUpdated();
+        }
+        this.disableConnectedStationMode();
+      }
+    }
 
     //update the stationElements array.
     this.stationElements.push(newStation);
@@ -340,17 +371,17 @@ export class MapService {
   /**
    * Calls the zoom() method a number of times equal to the zoomCount.
    *
-   * @param zoomOrigin The specific location on the canvas to zoom. Optional; defaults to the center of the canvas.
    * @param pinch Remove delay if zoom is a pinch zoom.
+   * @param zoomOrigin The specific location on the canvas to zoom. Optional; defaults to the center of the canvas.
    */
-  handleZoom(zoomOrigin = this.getCanvasCenterPoint(), pinch: boolean): void {
+  handleZoom(pinch: boolean, zoomOrigin = this.getCanvasCenterPoint()): void {
 
     const zoomLogic = () => {
       if (this.zoomCount$.value > 0) {
         this.zoom(true, zoomOrigin);
         this.zoomCount$.next(this.zoomCount$.value - 1);
         if (this.zoomCount$.value > 0) {
-          this.handleZoom(zoomOrigin, pinch);
+          this.handleZoom(pinch, zoomOrigin);
         }
       }
 
@@ -358,7 +389,7 @@ export class MapService {
         this.zoom(false, zoomOrigin);
         this.zoomCount$.next(this.zoomCount$.value + 1);
         if (this.zoomCount$.value < 0) {
-          this.handleZoom(zoomOrigin, pinch);
+          this.handleZoom(pinch, zoomOrigin);
         }
       }
     };
@@ -418,6 +449,152 @@ export class MapService {
   }
 
   /**
+   * Set a bounding box around the edge of the map.
+   *
+   * @returns A number representing how for out from the edge of the screen a box should be.
+   */
+  boundingBox(): number {
+    //Dynamically set the size of the bounding box based on screen size.
+    if (((window.innerHeight + window.innerWidth) / 2) * .01 < 30) {
+      //Set the size of the box based on screen size.
+      return Math.floor(((window.innerHeight + window.innerWidth) / 2) * .01);
+    } else {
+      //If a screen is above a certain size just return 30.
+      return 30;
+    }
+  }
+
+  /**
+   * Logic for finding top-left or bottom-right canvas or map points.
+   *
+   * @param pointType A mapPoint or a canvasPoint.
+   * @param isMax Is the point the top-left corner of the map or the bottom-right? Bottom-right is the max.
+   * @returns An object with the points.
+   */
+  private getEdgePoint(pointType: 'mapPoint' | 'canvasPoint', isMax: boolean): Point {
+    const orderedYPoints = this.stationElements.map((station) => station[pointType].y).sort((a, b) => a - b);
+    const orderedXPoints = this.stationElements.map((station) => station[pointType].x).sort((a, b) => a - b);
+
+    const x = isMax ? orderedXPoints[orderedXPoints.length - 1] + STATION_WIDTH : orderedXPoints[0];
+    const y = isMax ? orderedYPoints[orderedYPoints.length - 1] + STATION_HEIGHT : orderedYPoints[0];
+
+    return {
+      x: x,
+      y: y
+    };
+  }
+
+  /**
+   * Gets the top-left mapPoint.
+   *
+   * @returns A point.
+   */
+  getMinMapPoint(): Point {
+    return this.getEdgePoint('mapPoint', false);
+  }
+
+  /**
+   * Gets the bottom-left mapPoint.
+   *
+   * @returns A point.
+   */
+  getMaxMapPoint(): Point {
+    return this.getEdgePoint('mapPoint', true);
+  }
+
+  /**
+   * Gets the top-left canvasPoint.
+   *
+   * @returns A point.
+   */
+  getMinCanvasPoint(): Point {
+    return this.getEdgePoint('canvasPoint', false);
+  }
+
+  /**
+   * Gets the bottom-right canvasPoint.
+   *
+   * @returns A point.
+   */
+  getMaxCanvasPoint(): Point {
+    return this.getEdgePoint('canvasPoint', true);
+  }
+
+  /**
+   * Scales the map when center method is called to allow as many stations as possible to be visible.
+   *
+   * @param onInit Determines if this is called during mapCanvas init.
+   */
+  private centerScale(onInit = false): void {
+    if (!this.canvasContext) {
+      throw new Error('Cannot get center point of canvas when canvas context is not set');
+    }
+    const canvasBoundingRect = this.canvasContext.canvas.getBoundingClientRect();
+
+    //We use the canvas points of each station here.
+    const minPoint = this.getMinCanvasPoint();
+    const maxPoint = this.getMaxCanvasPoint();
+
+    //Zooming in and zooming out need to have different sized bounding boxes to work.
+    const zoomInBox = this.boundingBox() + CENTER_ZOOM_BUFFER;
+    const zoomOutBox = this.boundingBox() - CENTER_ZOOM_BUFFER;
+
+    //Zoom in.
+    if ((zoomInBox < minPoint.y
+      && canvasBoundingRect.height - zoomInBox > maxPoint.y + STATION_HEIGHT
+      && canvasBoundingRect.width - zoomInBox > maxPoint.x + STATION_WIDTH
+      && zoomInBox < minPoint.y
+      && this.mapScale$.value < MAX_SCALE) || this.mapScale$.value < SCALE_REDUCED_RENDER
+    ) {
+      this.zoomCount$.next(this.zoomCount$.value + 1);
+      this.handleZoom(onInit);
+      this.setCenterPanAndScale(onInit);
+      //Zoom out.
+    } else if ((zoomOutBox > minPoint.y
+      || canvasBoundingRect.height - zoomOutBox < maxPoint.y + STATION_HEIGHT
+      || canvasBoundingRect.width - zoomOutBox < maxPoint.x + STATION_WIDTH
+      || zoomOutBox > minPoint.y) && this.mapScale$.value > SCALE_REDUCED_RENDER/ZOOM_VELOCITY
+    ) {
+      this.zoomCount$.next(this.zoomCount$.value - 1);
+      this.handleZoom(onInit);
+      this.setCenterPanAndScale(onInit);
+    }
+  }
+
+  /**
+   * Smoothly sets the scale and pans the map to center.
+   *
+   * @param onInit Determines if this is called during mapCanvas init.
+   */
+  setCenterPanAndScale(onInit: boolean): void {
+    if (onInit) {
+      this.centerScale(onInit);
+      //TODO: set canvaspoint to map center.
+    } else {
+      setTimeout(() => {
+        this.centerScale(onInit);
+        //TODO: incrementally pan map to center.
+      }, 4);
+    }
+  }
+
+  /**
+   * Centers the map by changing this.currentCanvasPoint to the map center point.
+   *
+   * @param onInit Determines if this is called during mapCanvas init.
+   */
+  center(onInit = false): void {
+    let adjustedCenter = this.getMapCenterPoint();
+    const canvasCenter = this.getCanvasCenterPoint();
+    adjustedCenter = {
+      x: adjustedCenter.x - canvasCenter.x / this.mapScale$.value,
+      y: adjustedCenter.y - canvasCenter.y / this.mapScale$.value
+    };
+    this.currentCanvasPoint$.next(adjustedCenter);
+    this.setCenterPanAndScale(onInit);
+  }
+
+  /**
    * Gets the center point of the canvas.
    *
    * @returns The center point of the canvas.
@@ -467,6 +644,20 @@ export class MapService {
   }
 
   /**
+   * Find's the center of a user's map by looking at the mapPoints of the furthest left, top, right and bottom stations.
+   *
+   * @returns The center point of the map.
+   */
+  getMapCenterPoint(): Point {
+    //We use the map points of each station here.
+    const minPoint = this.getMinMapPoint();
+    const maxPoint = this.getMaxMapPoint();
+
+    const center: Point = { x: Math.floor((minPoint.x + maxPoint.x) / 2), y: Math.floor((minPoint.y + maxPoint.y) / 2) };
+    return center;
+  }
+
+  /**
    * Gets the x-coordinate on the map for a given canvas x-coordinate.
    *
    * @param canvasX The x-coordinate on the canvas.
@@ -497,6 +688,16 @@ export class MapService {
       x: this.getMapX(canvasPoint.x),
       y: this.getMapY(canvasPoint.y)
     };
+  }
+
+  /**
+   * Set's isAddingConnected property of station to false if it's true.
+   */
+  disableConnectedStationMode(): void {
+    this.stationElements.filter(station => station.isAddingConnected)
+    .map(connectedStation => {
+      connectedStation.isAddingConnected = false;
+    });
   }
 
   /**
@@ -565,6 +766,15 @@ export class MapService {
         console.error(`No flows contain the flow: ${flow.title} ${flow.rithmId}`);
       }
     }
+  }
+
+  /**
+   * Disable publish button until some changes in map/station.
+   *
+   * @returns Returns true if no stations are updated and false if any station is updated.
+   */
+  get mapHasChanges(): boolean {
+    return this.stationElements.some((station) => station.status !== MapItemStatus.Normal);
   }
 
 }
