@@ -1,8 +1,12 @@
-import { Component, EventEmitter, forwardRef, Input, OnInit, Output } from '@angular/core';
+import { Component, EventEmitter, forwardRef, Input, OnDestroy, OnInit, Output } from '@angular/core';
 // eslint-disable-next-line max-len
 import { ControlValueAccessor, FormBuilder, FormGroup, NG_VALIDATORS, NG_VALUE_ACCESSOR, ValidationErrors, Validator } from '@angular/forms';
 import { MatCheckboxChange } from '@angular/material/checkbox';
+import { Subject } from 'rxjs';
+import { takeUntil } from 'rxjs/operators';
+import { StationService } from 'src/app/core/station.service';
 import { Question, QuestionFieldType } from 'src/models';
+
 /**
  * Station Field Component.
  */
@@ -23,7 +27,7 @@ import { Question, QuestionFieldType } from 'src/models';
     }
   ]
 })
-export class StationFieldComponent implements OnInit, ControlValueAccessor, Validator {
+export class StationFieldComponent implements OnInit, ControlValueAccessor, Validator, OnDestroy {
 
   /** The document field to display. */
   @Input() field!: Question;
@@ -48,7 +52,7 @@ export class StationFieldComponent implements OnInit, ControlValueAccessor, Vali
 
   /** Instruction field to display. */
   instructionField: Question = {
-    rithmId: '3j4k-3h2j-hj4j',
+    rithmId: '',
     prompt: 'Instructions',
     questionType: QuestionFieldType.Instructions,
     isReadOnly: false,
@@ -59,7 +63,7 @@ export class StationFieldComponent implements OnInit, ControlValueAccessor, Vali
 
   /** Label field to display. */
   labelField: Question = {
-    rithmId: '3j4k-3h2j-hj4j',
+    rithmId: '',
     prompt: 'Name your field',
     questionType: QuestionFieldType.ShortText,
     isReadOnly: false,
@@ -71,37 +75,49 @@ export class StationFieldComponent implements OnInit, ControlValueAccessor, Vali
 
   /** The field for adding an option to a select field. */
   selectOptionField: Question = {
-    rithmId: '3j4k-3h2j-hj4j',
+    rithmId: '',
     prompt: 'Add Option',
     questionType: QuestionFieldType.ShortText,
     isReadOnly: false,
     isRequired: true,
     isPrivate: false,
     children: [],
+    isPossibleAnswer: true,
   };
 
   /** The field for adding an item to a checklist field. */
   checklistOptionField: Question = {
-    rithmId: '3j4k-3h2j-hj4j',
+    rithmId: '',
     prompt: 'Add Item',
     questionType: QuestionFieldType.ShortText,
     isReadOnly: false,
     isRequired: true,
     isPrivate: false,
     children: [],
+    isPossibleAnswer: true,
   };
 
   /** Array of options for a select/multi-select/checklist field. */
   options: Question[] = [];
 
+  /** The RithmId of the Station. */
+  @Input() stationRithmId = '';
+
+  /** Observable for when the component is destroyed. */
+  private destroyed$ = new Subject<void>();
+
   constructor(
     private fb: FormBuilder,
+    private stationService: StationService,
   ) { }
 
   /**
    * On component initialization.
    */
   ngOnInit(): void {
+    this.instructionField.rithmId = this.field.rithmId;
+    this.instructionField.value = this.field.prompt;
+    this.labelField.rithmId = this.field.rithmId;
     this.labelField.value = this.field.prompt;
     this.labelField.questionType = this.field.questionType;
     if (this.field.questionType === this.fieldType.Select
@@ -109,12 +125,28 @@ export class StationFieldComponent implements OnInit, ControlValueAccessor, Vali
       || this.field.questionType === this.fieldType.CheckList) {
       this.addOption(this.field.questionType);
     }
-
     this.stationFieldForm = this.fb.group({
       instructionsField: [''],
       [this.field.questionType]: [''],
       optionField: ['']
     });
+    this.stationFieldForm.valueChanges.pipe(takeUntil(this.destroyed$))
+      .subscribe(() => {
+       this.stationService.touchStationForm();
+      });
+  }
+
+  /**
+   * Returns the appropiate label tag.
+   *
+   * @returns The Label tag for each additional field.
+   */
+   get labelTag(): string{
+    // eslint-disable-next-line max-len
+    const label = this.field.questionType === this.fieldType.Select ? 'Add Option'
+    : this.field.questionType === this.fieldType.MultiSelect || this.field.questionType === this.fieldType.CheckList ? 'Add Item'
+    : 'Name your field';
+    return label;
   }
 
   /**
@@ -140,6 +172,8 @@ export class StationFieldComponent implements OnInit, ControlValueAccessor, Vali
    * @param fieldType The field type.
    */
   addOption(fieldType: QuestionFieldType): void {
+    this.selectOptionField.rithmId = this.field.rithmId;
+    this.checklistOptionField.rithmId = this.field.rithmId;
     this.options.push(fieldType === QuestionFieldType.Select ? this.selectOptionField : this.checklistOptionField);
   }
 
@@ -155,10 +189,11 @@ export class StationFieldComponent implements OnInit, ControlValueAccessor, Vali
   /**
    * Sets the required status of a field.
    *
-   * @param ob Observes MatCheckbox changes.
+   * @param checkboxEvent Event fired when the checkbox changes.
    */
-  setRequired(ob: MatCheckboxChange): void {
-    this.field.isRequired = ob.checked;
+  setRequired(checkboxEvent: MatCheckboxChange): void {
+    this.field.isRequired = checkboxEvent.checked;
+    this.stationService.touchStationForm();
   }
 
   /**
@@ -168,6 +203,20 @@ export class StationFieldComponent implements OnInit, ControlValueAccessor, Vali
    */
   setPrivate(checkboxEvent: MatCheckboxChange): void {
     this.field.isPrivate = checkboxEvent.checked;
+    this.stationService.touchStationForm();
+  }
+
+  /**
+   * Sets the read-only status of a field.
+   *
+   * @param checkboxEvent Event fired when the checkbox changes.
+   */
+  setEditable(checkboxEvent: MatCheckboxChange): void {
+    this.field.isReadOnly = checkboxEvent.checked;
+    if (!this.field.isReadOnly) {
+       this.field.isRequired = false;
+    }
+    this.stationService.touchStationForm();
   }
 
   /**
@@ -228,5 +277,13 @@ export class StationFieldComponent implements OnInit, ControlValueAccessor, Vali
         message: 'Station field is invalid'
       }
     };
+  }
+
+  /**
+   * Completes all subscriptions.
+   */
+  ngOnDestroy(): void {
+    this.destroyed$.next();
+    this.destroyed$.complete();
   }
 }
