@@ -1,4 +1,4 @@
-import { QuestionFieldType } from '.';
+import { QuestionFieldType, QuestionAnswer } from '.';
 import { PossibleAnswer } from './possible-answer';
 
 /**
@@ -35,4 +35,12 @@ export interface Question {
   /** Value of the new field (what do you call your field?). */
   value?: string;
 
+  /** Value of the Rithm ID current station to identify previous field. */
+  originalStationRithmId?: string | null;
+
+  /** Whether the added item is a possibleAnswer (FrontEndOnly). */
+  isPossibleAnswer?: boolean;
+
+  /** The answers to each question.*/
+  answer?: QuestionAnswer;
 }
