@@ -1,14 +1,12 @@
 const colors = require('./rithm-colors');
 
 module.exports = {
-  purge: {
-    enabled: process.env.NODE_ENV === 'production' || process.env.NODE_ENV === 'test',
-    content: [
-      './src/**/*.{html,ts}'
-    ]
-  },
+  // mode: 'jit',
+  content: [
+    './src/**/*.{html,ts}'
+  ],
   presets: [],
-  darkMode: false, // or 'media' or 'class'
+  darkMode: 'class', // or 'media' or 'class'
   theme: {
     screens: {
       sm: '640px',
